idt: Fix handler signatures
Note that the signatures are not checked by Rust. If the handler
signature is invalid, LLVM will not generate a correct epilogue, leading
 to a general protection fault.

It would be nice to be able to check these in some way, maybe with
macros? I tried a type safe IDT earlier, like in the x86-64 crate, but
it's a massive PITA, since we have to specify each field individually as
with/without error for the signature. Interrupts handlers can
have an error code or not, which changes the stack frame.